# backend/internal/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.createServiceDatabaseRequest:
    properties:
      app_ref:
        type: string
      backup_enabled:
        type: boolean
      backup_retention:
        type: string
      backup_schedule:
        type: string
      database:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Platform API to order and manage cloud resources — managed databases, apps, VMs (app servers), AI models, and public API endpoints — on the DADA Cloud platform.
    Mutations are asynchronous: create/update/delete enqueue an operation and return 202 Accepted with an operation object; poll GET /projects/{projectId}/operations/{operationId} until the operation reaches a terminal status (Ready or Failed).
  title: DADA Cloud Console API
  version: "1.0"
paths:
...
  /projects/{projectId}/environments/{envId}/databases:
    get:
      description: Returns all managed PostgreSQL (ServiceDatabaseV2) resources in
        the given project environment, with their live phase/status.
      operationId: listDatabases
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Environment UUID
        in: path
        name: envId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object with a databases array of ResourceSnapshot
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List databases in an environment
      tags:
      - database
    post:
      consumes:
      - application/json
      description: 'Provisions a new managed PostgreSQL database (ServiceDatabaseV2).
        app_ref is optional: omit it for a standalone, environment-level database,
        or set it to bind the database to an app''s chart. Asynchronous: returns 202
        with an operation; poll the operation until it reaches a terminal status.'
      operationId: createDatabase
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Environment UUID
        in: path
        name: envId
        required: true
        type: string
      - description: Database specification
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createServiceDatabaseRequest'
      produces:
      - application/json
      responses:
        "202":
          description: object with the accepted operation
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Order a managed PostgreSQL database
      tags:
      - database
  /projects/{projectId}/environments/{envId}/databases/{name}:
    delete:
      description: 'Destructive: permanently removes a managed PostgreSQL database
        (ServiceDatabaseV2) and its data. The agent drops the CR entry from git and
        Argo prunes it. Asynchronous: returns 202 with an operation; poll the operation
        until terminal. Only k8s (Crossplane) databases are deletable here; VM-hosted
        databases are managed as apps.'
      operationId: deleteDatabase
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Environment UUID
        in: path
        name: envId
        required: true
        type: string
      - description: Database resource name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: object with the accepted operation
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a managed PostgreSQL database
      tags:
      - database
...
  /projects/{projectId}/environments/{envId}/databases/{name}/credentials:
    get:
      description: Reveals the host, port, database name, username and password for
        a managed PostgreSQL database (ServiceDatabaseV2) by reading its Crossplane
        connection secret. Requires reveal=true and write access; every reveal is
        audited. Returns 404 while the database is still provisioning (no secret yet)
        and 503 when in-cluster credential access is not configured.
      operationId: getDatabaseCredentials
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Environment UUID
        in: path
        name: envId
        required: true
        type: string
      - description: Database resource name
        in: path
        name: name
        required: true
        type: string
      - description: Must be true to reveal the credentials
        in: query
        name: reveal
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: object with host, port, database, username, password
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reveal a database's connection credentials
      tags:
      - database
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"